requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
gunicorn==20.1.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.2.0
Pillow==9.3.0
